Document Parser 3
This is the 3rd and final version of Document Parser. Here we allow the
user to give keywords. From there we use functions to search through
those keywords in several texts taken from Project Gutenberg. The end
result spits out the Title, Author, Translator (if included),
Illustrator (if included) and how many of each word is in each text.